Store `darkMode` value in LocalStorage
This commit adds a Recoil effect which checks localStorage for a `darkMode`
item. If it exists, the value of the `darkModeState` atom is updated with the
users stored value.

If it's not stored, the localStorage item is set with the `darkModeState`'s
current value.

By doing this, the app is able to preserve a users Dark Mode preference and it
will be preserved each time they come back to the App.